organized file tree for services folder
