+++ bright-ideas/src/app/services/follower-endpoint/follower-endpoint.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse} from '@angular/common/http';
 import { Observable, Observer } from 'rxjs';
-import { Follower } from '../models/follower';
+
 
 
 @Injectable({ providedIn: 'root' })
