+++ bright-ideas/src/app/services/media-endpoint/media-endpoint.service.ts
@@ -1,31 +1,20 @@
 import { Injectable } from '@angular/core';
 import { Observable, Observer } from 'rxjs';
 import { HttpClient, HttpErrorResponse} from '@angular/common/http';
-import { Message } from '../models/message';
+import { Media } from '../../models/media';
 
 
 @Injectable({
   providedIn: 'root'
 })
-export class MessageEndpointService {
+export class MediaEndpointService {
   private url = 'https://bright-ideas-api.herokuapp.com';
 
   constructor(private http: HttpClient) { }
 
-  getAllMessages(): Observable<any[]> {
-    return Observable.create((observer: Observer<any[]>) => {
-       this.http.get(`${this.url}/messages`).subscribe((res: any[]) => {
-          observer.next(res);
-          observer.complete();
-       },
-       (error: HttpErrorResponse) => {
-          observer.error(error);
-       });
-    });
-  }
-  getMessagebyId(id: string): Observable<any> {
-    return Observable.create((observer: Observer<any>) => {
-       this.http.get(`${this.url}/message/${id}`).subscribe((res: any) => {
+  getAllMedia(): Observable<Media[]> {
+    return Observable.create((observer: Observer<Media[]>) => {
+       this.http.get(`${this.url}/media`).subscribe((res: Media[]) => {
           observer.next(res);
           observer.complete();
        },
@@ -35,9 +24,17 @@ export class MessageEndpointService {
     });
   }
 
-
-
-
+  getMediabyId(id: string): Observable<Media> {
+    return Observable.create((observer: Observer<Media>) => {
+       this.http.get(`${this.url}/media/${id}`).subscribe((res: Media) => {
+          observer.next(res);
+          observer.complete();
+       },
+       (error: HttpErrorResponse) => {
+          observer.error(error);
+       });
+    });
+  }
 
 
 }
