+++ bright-ideas/src/app/services/media-endpoint/media-queries.js
