+++ bright-ideas/src/app/services/update-endpoint/update-queries.js
